Correct dependency name in metadata
The depends field of the library.properties metadata file specifies the dependencies that should be installed along with this library via Library Manager. This requires the use of the library name that is locked in when a library is submitted to Library Manager.

The "LiquidCrystal I2C" was named that way when it was added to Library Manager. Since then, the library maintainer changed the name in the metadata, but this does not affect the name in Library Manager (all the releases made under the new name are rejected by the indexer).